requirements.txt
# Environment Variables
python-dotenv==1.0.0

# Flask Core
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==6.0.1
Flask-Bcrypt==1.0.1
flasgger==0.9.7.1

# Production Server
gunicorn==23.0.0

# Database
SQLAlchemy==2.0.43
psycopg2-binary==2.9.9

# Cryptography & Security
cryptography==46.0.1
PyJWT==2.10.1
passlib==1.7.4
bcrypt==5.0.0
argon2-cffi==23.1.0

# Two-Factor Authentication
qrcode==8.2
Pillow==11.3.0

# Web Server
Werkzeug==3.1.3

# Utilities
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3

# Type Checking & Dependencies
typing-extensions==4.15.0
importlib-metadata==8.7.0
zipp==3.23.0

# Binary Dependencies
cffi==2.0.0
pycparser==2.23
greenlet==3.2.4
